refactor: rename 'feriado' helper to 'feriados'
Rename made to make names of helper and test compatible
